+++ data/users.json
@@ -6,7 +6,7 @@
         "password": "202cb962ac59075b964b07152d234b70",
         "firstName": "Darth",
         "lastName": "Vader",
-        "session": "a9d87fad94a20d66bfc40ffe10567b09",
+        "session": "a2d460aa1826b85a4eabfb328bd284a8",
         "role": "user"
     },
     {
