+++ index.js
@@ -1,15 +1,11 @@
 const express = require('express');
 const cors = require('cors');
-const md5 = require('md5');
 const cookieParser = require('cookie-parser');
-const {
-    login,
-    registerUser,
-    updateUser,
-    logout,
-    findLoggedInUser,
-} = require('./businessRules/users');
-const { getUsers, getIdeas: getIdeasUtil } = require('./utils/storage');
+
+const loginRoute = require('./routes/login');
+
+const { registerUser, updateUser, logout } = require('./businessRules/users');
+const { getUsers } = require('./utils/storage');
 const { deleteUser, getUser } = require('./models/users');
 const {
     getTotalSumDonatedForIdea,
@@ -17,7 +13,7 @@ const {
     createDonationByRegisteredUser,
 } = require('./businessRules/donations');
 const bodyParser = require('body-parser');
-const { updateIdea, deleteIdea, getIdeas, getIdea } = require('./models/ideas');
+const { updateIdea, deleteIdea, getIdeas } = require('./models/ideas');
 const {
     sortIdeasByTotalDonationSum,
     getPendingIdeas,
@@ -184,37 +180,7 @@ app.post('/donations', (req, res) => {
 });
 
 //Login
-
-app.post('/login', (req, res) => {
-    try {
-        const sessionId = login(req.body.username, md5(req.body.password));
-
-        res.cookie('userLoginSession', sessionId, {
-            sameSite: 'None',
-            secure: true,
-            httpOnly: true,
-            maxAge: 900000,
-        });
-    } catch (error) {
-        res.status(400).send(error.message);
-    }
-
-    res.status(200).send('Successfully logged in.');
-});
-
-app.get('/login', (req, res) => {
-    try {
-        const user = findLoggedInUser(req.body.userLoginSession);
-        res.status(200).json({
-            status: 'OK',
-            message: 'You are Logged in.',
-            name: user.firstName,
-            role: user.role,
-        });
-    } catch (error) {
-        res.status(404).send(error.message);
-    }
-});
+app.use('/login', loginRoute);
 
 //Logout
 app.post('/logout', (req, res) => {
